+++ geopandas/io/tests/generate_legacy_storage_files.py
@@ -45,7 +45,7 @@ def create_pickle_data():
         crs="EPSG:4326",
     )
 
-    return dict(gdf_the_geom=gdf_the_geom, gdf_crs=gdf_crs)
+    return {"gdf_the_geom": gdf_the_geom, "gdf_crs": gdf_crs}
 
 
 def platform_name():
@@ -80,7 +80,7 @@ def write_legacy_pickles(output_dir):
 
 def main():
     if len(sys.argv) != 3:
-        exit(
+        sys.exit(
             "Specify output directory and storage type: generate_legacy_"
             "storage_files.py <output_dir> <storage_type> "
         )
@@ -91,7 +91,7 @@ def main():
     if storage_type == "pickle":
         write_legacy_pickles(output_dir=output_dir)
     else:
-        exit("storage_type must be one of {'pickle'}")
+        sys.exit("storage_type must be one of {'pickle'}")
 
 
 if __name__ == "__main__":